+++ src/App.tsx
@@ -1,6 +1,7 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import Main from './pages/Main';
 import Login from './pages/Login';
+import List from './pages/List';
 
 function App() {
   return (
@@ -10,6 +11,9 @@ function App() {
           <Route index element={<Main />} />
           <Route path="login">
             <Route index element={<Login />} />
+          </Route>
+          <Route path="list">
+            <Route index element={<List />} />
           </Route>
         </Route>
       </Routes>
